res/values-sv/strings.xml - Swedish Translate
Here is a Swedish translate. It is my first, and I hope that I'm doing it right ;)
Btw, it's a little bit hard to translate when I can't see the big picture, but I did my best, and if there some wrong translated word maybe I can change it later?